# Make numpy arrow extractor faster