--- a/src/datasets/features.py
+++ b/src/datasets/features.py
@@ -379,7 +379,7 @@ class ArrayExtensionArray(pa.ExtensionArray):
     def __getitem__(self, i):
         return self.storage[i]
 
-    def to_numpy(self):
+    def to_numpy(self, zero_copy_only=True):
         storage: pa.ListArray = self.storage
         size = 1
         for i in range(self.type.ndims):
@@ -390,7 +390,7 @@ class ArrayExtensionArray(pa.ExtensionArray):
         # https://github.com/wesm/arrow/blob/c07b9b48cf3e0bbbab493992a492ae47e5b04cad/python/pyarrow/types.pxi#L821
         # see https://arrow.apache.org/docs/python/generated/pyarrow.Array.html#pyarrow.Array.to_numpy
         # and https://issues.apache.org/jira/browse/ARROW-2871?jql=text%20~%20%22boolean%20to_numpy%22
-        zero_copy_only = is_primitive(storage.type) and not is_boolean(storage.type)
+        zero_copy_only = zero_copy_only and is_primitive(storage.type) and not is_boolean(storage.type)
         numpy_arr = storage.to_numpy(zero_copy_only=zero_copy_only)
         numpy_arr = numpy_arr.reshape(len(self), *self.type.shape)
         return numpy_arr