--- a/src/datasets/formatting/formatting.py
+++ b/src/datasets/formatting/formatting.py
@@ -148,19 +148,21 @@ class NumpyArrowExtractor(BaseArrowExtractor[dict, np.ndarray, dict]):
         return _unnest(self.extract_batch(pa_table))
 
     def extract_column(self, pa_table: pa.Table) -> np.ndarray:
-        series = pa_table.to_pandas(types_mapper=pandas_types_mapper)[pa_table.column_names[0]]
-        return self._series_to_numpy(series)
+        return self._arrow_array_to_numpy(pa_table[pa_table.column_names[0]])
 
     def extract_batch(self, pa_table: pa.Table) -> dict:
-        df = pa_table.to_pandas(types_mapper=pandas_types_mapper)
-        return {k: self._series_to_numpy(v) for k, v in df.items()}
+        return {col: self._arrow_array_to_numpy(pa_table[col]) for col in pa_table.column_names}
 
-    def _series_to_numpy(self, series: pd.Series) -> np.ndarray:
-        # to_numpy takes too much time for series of objects (like series of np.array)
-        if series.dtype == np.object:
-            array = series.tolist()
+    def _arrow_array_to_numpy(self, pa_array: pa.Array) -> np.ndarray:
+        if isinstance(pa_array, pa.ChunkedArray):
+            # don't call to_numpy() directly or we end up with a np.array with dtype object
+            # call to_numpy on the chunks instead
+            array: List[np.ndarray] = [
+                row for chunk in pa_array.chunks for row in chunk.to_numpy(zero_copy_only=False)
+            ]
         else:
-            array = series.to_numpy()
+            # cast to list of arrays or we end up with a np.array with dtype object
+            array: List[np.ndarray] = pa_array.to_numpy().tolist()
         return np.array(array, copy=False, **self.np_array_kwargs)